[codegen] Get new-style codegen to a checkpoint
Everything is now working that was previously working; the main
outstanding thing I want to do is write up some README files explaining
how the various parts fit together.
