+++ font-codegen/Cargo.toml
@@ -7,15 +7,12 @@ autotests = false
 publish = false
 default-run = "font-codegen"
 
-[dev-dependencies]
-trybuild = { version = "1.0.49", features = ["diff"] }
-zerocopy = "0.6"
-bitflags = "1.3"
-
 [dependencies]
 rustfmt-wrapper = "0.1"
 regex = "1.5"
 miette = { version =  "4.6", features = ["fancy"] }
-syn =  { version = "1.0.86", features = ["parsing", "full"] }
+syn =  { version = "1.0", features = ["parsing", "full"] }
 proc-macro2 =  { version = "1.0", features = ["span-locations"]}
 quote = "1.0"
+toml = "0.5"
+serde = {version = "1.0", features = ["derive"] }
