+++ font-tables/src/tables/head.rs
@@ -1,60 +1,18 @@
-use font_types::{BigEndian, Fixed, LongDateTime};
+//! The [head](https://docs.microsoft.com/en-us/typography/opentype/spec/head) table
 
-font_types::tables! {
-    /// <https://docs.microsoft.com/en-us/typography/opentype/spec/head>
-    Head {
-        /// Major version number of the font header table — set to 1.
-        major_version: BigEndian<u16>,
-        /// Minor version number of the font header table — set to 0.
-        minor_version: BigEndian<u16>,
-        /// Set by font manufacturer.
-        font_revision: BigEndian<Fixed>,
-        /// To compute: set it to 0, sum the entire font as uint32, then
-        /// store 0xB1B0AFBA - sum. If the font is used as a component in a
-        /// font collection file, the value of this field will be
-        /// invalidated by changes to the file structure and font table
-        /// directory, and must be ignored.
-        checksum_adjustment: BigEndian<u32>,
-        /// Set to 0x5F0F3CF5.
-        magic_number: BigEndian<u32>,
-        /// See the flags enum
-        flags: BigEndian<u16>,
-        /// Set to a value from 16 to 16384. Any value in this range is
-        /// valid. In fonts that have TrueType outlines, a power of 2 is
-        /// recommended as this allows performance optimizations in some
-        /// rasterizers.
-        units_per_em: BigEndian<u16>,
-        /// Number of seconds since 12:00 midnight that started January 1st
-        /// 1904 in GMT/UTC time zone.
-        created: BigEndian<LongDateTime>,
-        /// Number of seconds since 12:00 midnight that started January 1st
-        /// 1904 in GMT/UTC time zone.
-        modified: BigEndian<LongDateTime>,
-        /// Minimum x coordinate across all glyph bounding boxes.
-        x_min: BigEndian<i16>,
-        /// Minimum y coordinate across all glyph bounding boxes.
-        y_min: BigEndian<i16>,
-        /// Maximum x coordinate across all glyph bounding boxes.
-        x_max: BigEndian<i16>,
-        /// Maximum y coordinate across all glyph bounding boxes.
-        y_max: BigEndian<i16>,
-        /// see somewhere else
-        mac_style: BigEndian<u16>,
-        /// Smallest readable size in pixels.
-        lowest_rec_ppem: BigEndian<u16>,
-        /// Deprecated (Set to 2).
-        font_direction_hint: BigEndian<i16>,
-        /// 0 for short offsets (Offset16), 1 for long (Offset32).
-        index_to_loc_format: BigEndian<i16>,
-        /// 0 for current format.
-        glyph_data_format: BigEndian<i16>,
-    }
-}
+#[path = "../../generated/generated_head.rs"]
+mod generated;
+
+use font_types::Tag;
+
+pub use generated::*;
+
+/// 'name'
+pub const TAG: Tag = Tag::new(b"head");
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use font_types::{test_helpers::BeBuffer, FontRead};
+    use font_types::{test_helpers::BeBuffer, Fixed, FontRead, LongDateTime};
 
     #[test]
     fn smoke_text() {
@@ -68,7 +26,7 @@ mod tests {
         buf.extend([0u16, 12]); // mac_style / ppem
         buf.extend([2i16, 1, 0]);
 
-        let head = Head::read(&buf).unwrap();
+        let head = super::Head::read(&buf).unwrap();
         assert_eq!(head.major_version(), 1);
         assert_eq!(head.minor_version(), 0);
         assert_eq!(head.font_revision(), Fixed::from_f64(2.8));
